third_party/WebKit/LayoutTests/inspector/tracing/timeline-window-filter.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script src="resources/timeline-data.js"></script>
<script>

function test()
{
    var timelineData = InspectorTest.timelineData();

    WebInspector.inspectorView.showPanel("timeline");
    var timeline = WebInspector.panels.timeline;
    var overviewPane = timeline._overviewPane;
    var model = timeline._model;

    function createFileReader(file, delegate)
    {
        return new InspectorTest.FakeFileReader(timelineData, delegate, model.saveToFile.bind(model));
    }

    InspectorTest.override(model, "_createFileReader", createFileReader);

    model.loadFromFile({}, new WebInspector.Progress());
    timeline._currentViews[0]._refresh();

    overviewPane.update();
    InspectorTest.addResult("OverviewPane:");
    dumpDividers(overviewPane._overviewCalculator);
    InspectorTest.addResult("");

    function dumpRecordsCountForRange(windowLeft, windowRight)
    {
        var view = timeline._currentViews[0];
        overviewPane._overviewGrid.setWindow(windowLeft, windowRight);
        view._refresh();
        InspectorTest.addResult("mode = " + timeline._overviewModeSetting.get());
        InspectorTest.addResult("range = " + windowLeft + " - " + windowRight);
        InspectorTest.addResult("time range = " + view._windowStartTime + " - " + view._windowEndTime);
        InspectorTest.addResult("records count: " + view._presentationModel.filteredRecords().length);
        dumpDividers(view._calculator);
        InspectorTest.addResult("");
    }

    function dumpFlameChartRecordsCountForRange(windowLeft, windowRight)
    {
        var mainView = timeline._currentViews[0]._mainView;
        overviewPane._overviewGrid.setWindow(windowLeft, windowRight);
        mainView.update();
        InspectorTest.addResult("mode = " + timeline._overviewModeSetting.get());
        InspectorTest.addResult("range = " + windowLeft + " - " + windowRight);
        InspectorTest.addResult("time range = " + mainView._timeWindowLeft + " - " + mainView._timeWindowRight);
        dumpDividers(mainView._calculator);
        InspectorTest.addResult("");
    }

    function dumpDividers(calculator)
    {
        var dividers = WebInspector.TimelineGrid.calculateDividerOffsets(calculator, 450).offsets;
        for (var i = 0; i < dividers.length; ++i)
            dividers[i] -= calculator.zeroTime();
        InspectorTest.addResult("divider offsets: [" + dividers.join(", ") + "]. We are expecting round numbers.");
    }

    timeline._overviewModeSetting.set(WebInspector.TimelinePanel.OverviewMode.Events);
    timeline._onModeChanged();
    dumpRecordsCountForRange(0, 1);
    dumpRecordsCountForRange(0.25, 0.75);
    dumpRecordsCountForRange(0.33, 0.66);

    timeline._overviewModeSetting.set(WebInspector.TimelinePanel.OverviewMode.Frames);
    timeline._onModeChanged();
    dumpRecordsCountForRange(0, 1);
    dumpRecordsCountForRange(0.25, 0.75);
    dumpRecordsCountForRange(0.33, 0.66);

    timeline._overviewModeSetting.set(WebInspector.TimelinePanel.OverviewMode.Frames);
    timeline._flameChartEnabledSetting.set(true);
    timeline._onModeChanged();
    // FIXME(394762): uncomment once we support loading tracing events.
//     dumpFlameChartRecordsCountForRange(0, 1);
//     dumpFlameChartRecordsCountForRange(0.25, 0.5);
//     dumpFlameChartRecordsCountForRange(0.33, 0.4);

    overviewPane._overviewGrid.setWindow(0.1, 0.9);

    InspectorTest.addResult("--------------------------------------------------------");
    InspectorTest.addResult("mode = " + timeline._overviewModeSetting.get());
    InspectorTest.addResult("time range = " + timeline._windowStartTime + " - " + timeline._windowEndTime);

    timeline._overviewModeSetting.set(WebInspector.TimelinePanel.OverviewMode.Events);
    timeline._flameChartEnabledSetting.set(false);
    timeline._onModeChanged();
    InspectorTest.addResult("");

    InspectorTest.addResult("mode = " + timeline._overviewModeSetting.get());
    InspectorTest.addResult("time range = " + timeline._windowStartTime + " - " + timeline._windowEndTime);
    timeline._overviewModeSetting.set(WebInspector.TimelinePanel.OverviewMode.Frames);
    timeline._onModeChanged();
    InspectorTest.addResult("");

    InspectorTest.addResult("mode = " + timeline._overviewModeSetting.get());
    InspectorTest.addResult("time range = " + timeline._windowStartTime + " - " + timeline._windowEndTime);
    timeline._overviewModeSetting.set(WebInspector.TimelinePanel.OverviewMode.Frames);
    timeline._flameChartEnabledSetting.set(true);
    timeline._onModeChanged();

    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline window filter.
</p>

<p>
It applies different ranges to the OverviewGrid and expects that current view reflects the change.
</p>
<p>
Also it detects that after switching to another view it gets the proper time range.
</p>

</body>
</html>
